added separate file dedicated to store types for the Card Display (cards and cards containers)to avoid dependency cycling
